(squash) klp-convert: flag annotations without relocations
Note: squash with ("livepatch: Add klp-convert tool")

Symbols found in the module_relocs KLP_SYMPOS annotation array that do
not have a corresponding rela symbol will remain unresolved by
klp-convert and throw the following warning:

  depmod: WARNING: /lib/modules/5.6.0+/kernel/lib/livepatch/test_klp_convert1.ko needs unknown symbol saved_command_line

This can happen if a symbol is specified by KLP_SYMPOS but not used
elsewhere in the code.  The macro defines an entry in module_relocs:

  #define KLP_SYMPOS(symbol, pos)                                         \
          {                                                               \
                  .sym = &symbol,                                         \
                  .sympos = pos,                                          \
          }

The entry and its rela are later removed by clear_sympos_annontations().
However, the symbol remains in the symbol table and undefined:

  % readelf --wide --symbols lib/livepatch/test_klp_convert1.ko
      ...
      54: 0000000000000000     0 NOTYPE  GLOBAL DEFAULT  UND saved_command_line

If there is no exported symbol of similar name, the module will refuse
to load.

Flag this as a fatal warning by adding another check to
sympos_sanity_check() to verify that all usr_symbols do indeed have a
relocation present.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>