(squash) klp-convert: suggestions from Marcos
Note: squash with ("livepatch: Add klp-convert tool")

> +static bool sympos_sanity_check(struct elf *klp_elf)
> ...
> +				WARN("Conflicting KLP_SYMPOS definition: %s.%s,%d vs. %s.%s,%d.",
> +				sp->object_name, sp->symbol_name, sp->pos,
> +				aux->object_name, aux->symbol_name, aux->pos);

The WARN message could be simplified by mentioning the different position,
something like:

	WARN("Conflicting KLP_SYMPOS definition: %s.%s %d vs. %d.",
	sp->object_name, sp->symbol_name, sp->pos aux->pos);

[ ... snip ... ]

> +static bool sympos_sanity_check(struct elf *klp_elf)
> ...
> +		if (!found_rela) {
> +			//sane = false;

At this point I believe that sane should be assigned to false to help the user
to know that the specified symbol isn't being used in the livepatch.

[ ... snip ... ]

> +static struct symbol_entry *find_sym_entry_by_name(char *name)
> ...
> +	if (found)
> +		return found;
> +
> +	return NULL;

Since found is either NULL or points to a symbol, the if condition can be
removed and return found directly.

[ ... snip ... ]

> +static bool find_sympos(struct symbol *s, struct sympos *sp)
> +{
> ...
> +	/* search symbol in symbols list */
> +	entry = find_sym_entry_by_name(s->name);
> +	if (entry) {
> +		sp->symbol_name = entry->symbol_name;
> +		sp->object_name = entry->object_name;

At this point I believe that it would be good to have a comment about sympos
being 0 means that the symbol wasn't specified by the user, so sympos 0 means
that the symbol is unique.

Suggested-by: Marcos Paulo de Souza <mpdesouza@suse.de>
Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>